PP-3373 - removing amount page and controllers
and putting the pieces into the index controller
