+++ app/controllers/adhoc_payment/get_index_controller.js
@@ -2,9 +2,14 @@
 
 // Node.js core dependencies
 const logger = require('winston')
+const currencyFormatter = require('currency-formatter')
 
 // Custom dependencies
 const response = require('../../utils/response').response
+
+function asGBP (amountInPence) {
+  return currencyFormatter.format((amountInPence / 100).toFixed(2), {code: 'GBP'})
+}
 
 module.exports = (req, res) => {
   const product = req.product
@@ -14,6 +19,15 @@ module.exports = (req, res) => {
     serviceName: product.serviceName,
     productName: product.name,
     productDescription: product.description
+  }
+
+  if (product.price) {
+    data.price = asGBP(product.price)
+  }
+  if (req.errorMessage) {
+    data.flash = {
+      genericError: req.errorMessage
+    }
   }
 
   logger.info(`[${correlationId}] initiating product payment for ${product.externalId}`)
