+++ app/controllers/adhoc_payment/index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-module.exports.index = require('./index_controller')
-module.exports.getAmount = require('./get_amount_controller')
-module.exports.submitAmount = require('./submit_amount_controller')
+module.exports.index = require('./get_index_controller')
+module.exports.postIndex = require('./post_index_controller')
 module.exports.confirmation = require('../payment_status_controller')
