+++ app/paths.js
@@ -14,9 +14,6 @@ module.exports = {
     success: '/successful',
     failure: '/failed'
   },
-  adhocPayment: {
-    amount: '/pay/:productExternalId/enter-amount'
-  },
   healthcheck: {
     path: '/healthcheck'
   },
